Align ValidatorInvoker specs with the repo spec conventions
Use NullLogger rather than substituting the logger - logging is not
specified - and split the fixture types into one file each.
