+++ Source/DotNET/Arc.Core.Specs/Validation/for_ValidatorInvoker/when_a_validator_rejects.cs
@@ -21,5 +21,5 @@ public class when_a_validator_rejects : given.a_validator_invoker
     [Fact] void should_keep_the_authored_messages() => _results.Select(_ => _.Message).ShouldContainOnly(RejectingValidator.NameMessage, RejectingValidator.EmailMessage);
     [Fact] void should_say_the_rejections_came_from_authored_rules() => _results.Select(_ => _.Reason).ShouldContainOnly(ValidationResultReason.Rule, ValidationResultReason.Rule);
     [Fact] void should_attribute_them_to_their_members() => _results.SelectMany(_ => _.Members).ShouldContainOnly("name", "email");
-    [Fact] void should_leave_the_authors_own_state_alone() => _results.Single(_ => _.Members.Contains("name")).State.ShouldEqual("author's state");
+    [Fact] void should_leave_the_authors_own_state_alone() => _results.Single(_ => _.Members.Contains("name")).State.ShouldEqual(RejectingValidator.AuthorState);
 }
